requirements.txt
# Olist E-Commerce Analysis Environment
# Python package requirements for marimo notebooks with DuckDB
# dbt for DuckDB
dbt-core>=1.10.0
dbt-duckdb>=1.10.0
dbt-duckdb[sources]>=1.10.0

# Fix protobuf compatibility with dbt
protobuf>=5.0,<7.0

# dbt Column Lineage
dbt-colibri>=0.2.7

# Code Quality & Linting Tools
sqlfluff>=3.5.0
sqlfluff-templater-dbt>=3.5.0
pre-commit>=4.4.0

# Development tools
pytest==7.4.3
black>=24.10.0
ruff==0.1.8
isort>=5.13.0

# Core Analysis Tools
marimo>=0.9.0
duckdb>=0.10.0
pandas>=2.0.0
polars>=0.20.0

# Visualization Libraries
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Data Processing & Utilities
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml==6.0.1
pyarrow>=12.0.0
requests>=2.31.0
tqdm>=4.65.0

# Optional: Additional Analysis Tools
scikit-learn>=1.3.0
scipy>=1.11.0

# Optional: Jupyter Support (if needed)
ipykernel>=6.25.0
ipywidgets>=8.1.0
